Create Concept of "Axis" for metrics
This commit creates the concept of an Axis, which is a group of Selectors which
have values in the same units. The Axis concept is needed for two reasons:

1. It allows us to properly label graphs that have multiple series.
1. It allows us to more efficiently batch queries: one query can grab all of the
data for multiple graphs.

The graph component now accepts an Executor (used to get data) and an Axis -
multiple charts may share the same Executor, but use different Axes.

Also included is the ability to provide a custom format function for an axis,
allowing "tick" values on the axis to be more readable.

This commit also significantly improves the titles and labels of charts - this
data is attached to the selectors and axes, and is used by components to improve
the display quality.